webapi/webapi-simd-nonw3c-tests/simd/float64x2.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>SIMD Test: float64x2</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://github.com/johnmccutchan/ecmascript_simd/blob/master/README.md">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

test(function () {
  assert_not_equals(undefined, SIMD.float64x2, "the type of float64x2 should not be undefined");
  assert_not_equals(undefined, SIMD.float64x2(1, 2), "the float64x2 object should not be undefined");
}, "Check if float64x2 object can be creat successful");

test(function () {
  var z1 = SIMD.float64x2.zero();
  assert_equals(0.0, z1.x, "the value of z1.x should be 0.0");
  assert_equals(0.0, z1.y, "the value of z1.y should be 0.0");
}, "Check if the zero function of float64x2 can change the parameter to 0.0");

test(function () {
  var z2 = SIMD.float64x2.splat(5.0);
  assert_equals(5.0, z2.x, "the value of z2.x should be 5.0");
  assert_equals(5.0, z2.y, "the value of z2.y should be 5.0");
}, "Check if the splat function of float64x2 can change the parameter to specified value");

test(function () {
  var m = SIMD.float32x4(9.0, 10.0, 11.0, 12.0);
  var nMask = SIMD.float64x2.fromFloat32x4Bits(m);
  var n = SIMD.float32x4.fromFloat64x2Bits(nMask);
  assert_equals(9.0, n.x, "the value of n.x should be 9.0");
  assert_equals(10.0, n.y, "the value of n.y should be 10.0");
  assert_equals(11.0, n.z, "the value of n.z should be 11.0");
  assert_equals(12.0, n.w, "the value of n.w should be 12.0");
}, "Check if the fromFloat32x4Bits method of float64x2 is valid");

test(function () {
  var m = SIMD.int32x4(0x3F800000, 0x40000000, 0x40400000, 0x40800000);
  var nMask = SIMD.float64x2.fromInt32x4Bits(m);
  var n = SIMD.int32x4.fromFloat64x2Bits(nMask);
  assert_equals(0x3F800000, n.x, "the value of n.x should be 0x3F800000");
  assert_equals(0x40000000, n.y, "the value of n.y should be 0x40000000");
  assert_equals(0x40400000, n.z, "the value of n.z should be 0x40400000");
  assert_equals(0x40800000, n.w, "the value of n.w should be 0x40800000");
}, "Check if the fromInt32x4Bits method of float64x2 is valid");

test(function () {
  var m = SIMD.int32x4(0x3F800000, 0x40000000, 0x40400000, 0x40800000);
  var nMask = SIMD.float64x2.fromInt32x4(m);
  var n = SIMD.int32x4.fromFloat64x2(nMask);
  assert_equals(0x3F800000, n.x, "the value of n.x should be 0x3F800000");
  assert_equals(0x40000000, n.y, "the value of n.y should be 0x40000000");
  assert_equals(0, n.z, "the value of n.z should be 0");
  assert_equals(0, n.w, "the value of n.w should be 0");
}, "Check if the fromInt32x4 method of float64x2 is valid");

test(function () {
  var m = SIMD.float32x4(9.0, 10.0, 11.0, 12.0);
  var nMask = SIMD.float64x2.fromFloat32x4(m);
  var n = SIMD.float32x4.fromFloat64x2(nMask);
  assert_equals(9.0, n.x, "the value of n.x should be 9.0");
  assert_equals(10.0, n.y, "the value of n.y should be 10.0");
  assert_equals(0, n.z, "the value of n.z should be 0");
  assert_equals(0, n.w, "the value of n.w should be 0");
}, "Check if the fromFloat32x4 method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(-4.0, -3.0);
  var c = SIMD.float64x2.abs(a);
  assert_equals(4.0, c.x, "the value of c.x should be 4.0");
  assert_equals(3.0, c.y, "the value of c.y should be 3.0");
}, "Check if the abs method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(-4.0, -3.0, -2.0, -1.0);
  var c = SIMD.float64x2.neg(a);
  assert_equals(4.0, c.x, "the value of c.x should be 4.0");
  assert_equals(3.0, c.y, "the value of c.y should be 3.0");
}, "Check if the neg method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(4.0, 3.0);
  var b = SIMD.float64x2(10.0, 20.0);
  var c = SIMD.float64x2.add(a, b);
  assert_equals(14.0, c.x, "the value of c.x should be 14.0");
  assert_equals(23.0, c.y, "the value of c.y should be 23.0");
}, "Check if the add method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(4.0, 3.0);
  var b = SIMD.float64x2(10.0, 20.0);
  var c = SIMD.float64x2.sub(a, b);
  assert_equals(-6.0, c.x, "the value of c.x should be -6.0");
  assert_equals(-17.0, c.y, "the value of c.y should be -17.0");
}, "Check if the sub method is valid");

test(function () {
  var a = SIMD.float64x2(4.0, 3.0);
  var b = SIMD.float64x2(10.0, 20.0);
  var c = SIMD.float64x2.mul(a, b);
  assert_equals(40.0, c.x, "the value of c.x should be 40.0");
  assert_equals(60.0, c.y, "the value of c.y should be 60.0");
}, "Check if the mul method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(4.0, 9.0);
  var b = SIMD.float64x2(2.0, 3.0);
  var c = SIMD.float64x2.div(a, b);
  assert_equals(2.0, c.x, "the value of c.x should be 2.0");
  assert_equals(3.0, c.y, "the value of c.y should be 3.0");
}, "Check if the div method is valid");

test(function () {
  var a = SIMD.float64x2(-20.0, 10.0);
  var lower = SIMD.float64x2(2.0, 1.0);
  var upper = SIMD.float64x2(2.5, 5.0);
  var c = SIMD.float32x4.clamp(a, lower, upper);
  assert_equals(2.0, c.x, "the value of c.x should be 2.0");
  assert_equals(5.0, c.y, "the value of c.y should be 5.0");
}, "Check if the clamp method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(-20.0, 10.0, 30.0, 0.5);
  var lower = SIMD.float64x2(2.0, 1.0, 50.0, 0.0);
  var c = SIMD.float64x2.min(a, lower);
  assert_equals(-20.0, c.x, "the value of c.x should be -20.0");
  assert_equals(1.0, c.y, "the value of c.y should be 1.0");
}, "Check if the min method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(-20.0, 10.0);
  var upper = SIMD.float64x2(2.5, 5.0);
  var c = SIMD.float64x2.max(a, upper);
  assert_equals(2.5, c.x, "the value of c.x should be 2.5");
  assert_equals(10.0, c.y, "the value of c.y should be 10.0");
}, "Check if the max method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(8.0, 4.0);
  var c = SIMD.float64x2.reciprocal(a);
  assert_equals(0.125, c.x, "the value of c.x should be 0.125");
  assert_equals(0.250, c.y, "the value of c.y should be 0.250");
}, "Check if the reciprocal method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(1.0, 0.25);
  var c = SIMD.float64x2.reciprocalSqrt(a);
  assert_equals(1.0, c.x, "the value of c.x should be 1.0");
  assert_equals(2.0, c.y, "the value of c.y should be 2.0");
}, "Check if the reciprocalSqrt method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(8.0, 4.0, 2.0, -2.0);
  var c = SIMD.float64x2.scale(a, 0.5);
  assert_equals(4.0, c.x, "the value of c.x should be 1.0");
  assert_equals(2.0, c.y, "the value of c.y should be 2.0");
}, "Check if the scale method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(16.0, 9.0);
  var c = SIMD.float64x2.sqrt(a);
  assert_equals(4.0, c.x, "the value of c.x should be 4.0");
  assert_equals(3.0, c.y, "the value of c.y should be 3.0");
}, "Check if the sqrt method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(4.0, 3.0);
  var xxxx = SIMD.float64x2.shuffle(a, SIMD.XXXX);
  assert_equals(4.0, xxxx.x, "the value of xxxx.x should be 4.0");
  assert_equals(4.0, xxxx.y, "the value of xxxx.y should be 4.0");
}, "Check if the shuffle  method of float64x2 is valid");

test(function () {
  var a    = SIMD.float64x2(1.0, 2.0);
  var b    = SIMD.float64x2(5.0, 6.0);
  var xyxy = SIMD.float64x2.shuffleMix(a, b, SIMD.XYXY);
  var zwzw = SIMD.float64x2.shuffleMix(a, b, SIMD.ZWZW);
  var xxxx = SIMD.float64x2.shuffleMix(a, b, SIMD.XXXX);
  assert_equals(1.0, xyxy.x, "the value of xyxy.x should be 1.0");
  assert_equals(5.0, xyxy.y, "the value of xyxy.y should be 5.0");
  assert_equals(1.0, zwzw.x, "the value of zwzw.x should be 1.0");
  assert_equals(6.0, zwzw.y, "the value of zwzw.y should be 6.0");
  assert_equals(1.0, xxxx.x, "the value of xxxx.x should be 1.0");
  assert_equals(5.0, xxxx.y, "the value of xxxx.y should be 5.0");
}, "Check if the shuffleMix  method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(16.0, 9.0);
  var c = SIMD.float64x2.withX(a, 20.0);
  assert_equals(20.0, c.x, "the value of c.x should be 20.0");
  assert_equals(9.0, c.y, "the value of c.y should be 9.0");
}, "Check if the withX method of float64x2 is valid");

test(function () {
  var a = SIMD.float64x2(16.0, 9.0);
  var c = SIMD.float64x2.withY(a, 20.0);
  assert_equals(16.0, c.x, "the value of c.x should be 16.0");
  assert_equals(20.0, c.y, "the value of c.y should be 20.0");
}, "Check if the withY method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.lessThan(m, n);
  assert_equals(-1, cmp.x, "the value of cmp.x should be -1");
  assert_equals(-1, cmp.y, "the value of cmp.y should be -1");
  assert_equals(0, cmp.z, "the value of cmp.z should be 0");
  assert_equals(0, cmp.w, "the value of cmp.w should be 0");
}, "Check if the lessThan method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.lessThanOrEqual(m, n);
  assert_equals(-1, cmp.x, "the value of cmp.x should be -1");
  assert_equals(-1, cmp.y, "the value of cmp.y should be -1");
  assert_equals(-1, cmp.z, "the value of cmp.z should be -1");
  assert_equals(-1, cmp.w, "the value of cmp.w should be -1");
}, "Check if the lessThanOrEqual method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.equal(m, n);
  assert_equals(0, cmp.x, "the value of cmp.x should be 0");
  assert_equals(0, cmp.y, "the value of cmp.y should be 0");
  assert_equals(-1, cmp.z, "the value of cmp.z should be -1");
  assert_equals(-1, cmp.w, "the value of cmp.w should be -1");
}, "Check if the equal method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.notEqual(m, n);
  assert_equals(-1, cmp.x, "the value of cmp.x should be -1");
  assert_equals(-1, cmp.y, "the value of cmp.y should be -1");
  assert_equals(0, cmp.z, "the value of cmp.z should be 0");
  assert_equals(0, cmp.w, "the value of cmp.w should be 0");
}, "Check if the notEqual method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.greaterThanOrEqual (m, n);
  assert_equals(0, cmp.x, "the value of cmp.x should be 0");
  assert_equals(0, cmp.y, "the value of cmp.y should be 0");
  assert_equals(-1, cmp.z, "the value of cmp.z should be -1");
  assert_equals(-1, cmp.w, "the value of cmp.w should be -1");
}, "Check if the greaterThanOrEqual  method of float64x2 is valid");

test(function () {
  var m = SIMD.float64x2(1.0, 2.0);
  var n = SIMD.float64x2(2.0, 2.0);
  var cmp;
  cmp = SIMD.float64x2.greaterThan (m, n);
  assert_equals(0, cmp.x, "the value of cmp.x should be 0");
  assert_equals(0, cmp.y, "the value of cmp.y should be 0");
  assert_equals(0, cmp.z, "the value of cmp.z should be 0");
  assert_equals(0, cmp.w, "the value of cmp.w should be 0");
}, "Check if the greaterThan  method of float64x2 is valid");

test(function () {
  var m = SIMD.int32x4.bool(true, true, false, false);
  var t = SIMD.float64x2(1.0, 2.0);
  var f = SIMD.float64x2(5.0, 6.0);
  var s = SIMD.float64x2.select(m, t, f);
  assert_equals(1.0, s.x, "the value of s.x should be 1");
  assert_equals(6.0, s.y, "the value of s.y should be 6");
}, "Check if the select method of float64x2 is valid");

test(function () {
  var t = SIMD.float64x2(1.0, 2.0);
  var a = SIMD.float64x2.bitsToInt32x4(t);
  var s = SIMD.int32x4.bitsToFloat64x2(a);
  assert_equals(1, s.x, "the value of s.x should be 1");
  assert_equals(2, s.y, "the value of s.y should be 2");
}, "Check if the bitsToInt32x4 method of float64x2 is valid");

</script>
